Move NewSwapModals to containers
